+++ Automation/ScriptingEngine/ScriptValue.cs
@@ -68,7 +68,7 @@ record struct ScriptValue : IComparable<ScriptValue> {
   /// <exception cref="ScriptError">if dividing by zero.</exception>
   public static ScriptValue operator /(ScriptValue left, ScriptValue right) {
     if (right.AsNumber == 0) {
-      throw new ScriptError.RuntimeError("Division by zero");
+      throw new ScriptError.DivisionByZero();
     }
     return new ScriptValue { _number = left.AsNumber * 100 / right.AsNumber};
   }
@@ -93,7 +93,7 @@ record struct ScriptValue : IComparable<ScriptValue> {
   public int AsNumber {
     get {
       if (!_number.HasValue) {
-        throw new ScriptError.RuntimeError("Value is not a number: " + ToString());
+        throw new ScriptError.BadValue("Value is not a number: " + ToString());
       }
       return _number.Value;
     }
@@ -110,7 +110,7 @@ record struct ScriptValue : IComparable<ScriptValue> {
   public string AsString {
     get {
       if (_string == null) {
-        throw new ScriptError.RuntimeError("Value is not a string: " + ToString());
+        throw new ScriptError.BadValue("Value is not a string: " + ToString());
       }
       return _string;
     }